requirements.txt
torch>=2.0.0
transformers>=4.30.0
torchaudio>=2.0.0
tokenizers>=0.13.0

# Data processing
numpy>=1.21.0
scipy>=1.9.0
librosa>=0.10.0
soundfile>=0.12.0
datasets>=2.0.0

# Configuration and utilities
pyyaml>=6.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# Training and evaluation
accelerate>=0.20.0
wandb>=0.15.0
tensorboard>=2.13.0
evaluate>=0.4.0
rouge-score>=0.1.2
sacrebleu>=2.3.0

# AWS and cloud
boto3>=1.26.0
sagemaker>=2.132.0
awscli>=1.27.0
s3fs>=2023.6.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.2.0
nbsphinx>=0.9.0

# Jupyter and notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.0.0

# Performance
torch-audiomentations>=0.11.0
torchdynamo>=2.0.0
triton>=2.0.0 